memory: fix file load
mloadf would not properly re-seek to the beginning of the file, thus
causing fread to load no bytes into memory. This fixes that error and
introduces a check to make sure that fread actually does what it's
supposed to do.

In addition, formatting on error message for a failing mmap was fixed.